Fixing test_tem.py
Due to an issue with Docker containerization, the master version of tests are exclusively used for pytest. So, in order for tests to pass on any branches, this tiny bug has to be fixed in master. I will discuss this issue with @fredericpoitevin and try to create a better solution going forward.